HOpenCV.CV.CxCore

Documentation

toFromIntegral :: (RealFrac c, Integral b, Integral a, Num b1) => (b1 -> c) -> a -> b

toFromIntegral2 :: (Integral a, Num b, Integral a1, Num b1, RealFrac a2, Integral b2) => (b -> b1 -> a2) -> a -> a1 -> b2

data CvSize

Constructors

CvSize

Fields

sizeWidth :: CInt

sizeHeight :: CInt

liftCvSize :: (RealFrac c, Num b) => (b -> c) -> CvSize -> CvSize

liftCvSize2 :: (Num b, Num b1, RealFrac a) => (b -> b1 -> a) -> CvSize -> CvSize -> CvSize

data CvRect

Constructors

CvRect

Fields

rectX :: CInt

rectY :: CInt

rectWidth :: CInt

rectHeight :: CInt

liftCvRect :: (RealFrac c, Num b) => (b -> c) -> CvRect -> CvRect

Apply a function to each component of a CvRect.

liftCvRect2 :: (Num b, Num b1, RealFrac a) => (b -> b1 -> a) -> CvRect -> CvRect -> CvRect

data Priv_IplImage

type IplImage = ForeignPtr Priv_IplImage

data Priv_CvMemStorage

type MemStorage = ForeignPtr Priv_CvMemStorage

data Priv_CvSeq a

type CvSeq a = ForeignPtr (Priv_CvSeq a)

newtype Depth

Constructors

Depth

Fields

unDepth :: CInt

iplDepth1u :: Depth

iplDepth8u :: Depth

iplDepth8s :: Depth

iplDepth16u :: Depth

iplDepth16s :: Depth

iplDepth32s :: Depth

validDepths :: [Depth]

iplDepth32f :: Depth

depthsLookupList :: [(CInt, Depth)]

iplDepth64f :: Depth

numToDepth :: CInt -> Maybe Depth

c_cvCreateMemStorage :: CInt -> IO (Ptr Priv_CvMemStorage)

cf_releaseMemStorage :: FunPtr (Ptr Priv_CvMemStorage -> IO ())

createMemStorage :: Int -> IO (ErrCV MemStorage)

cvFree :: FunPtr (Ptr a -> IO ())

c_cvCreateImage :: CInt -> CInt -> CInt -> CInt -> IO (Ptr Priv_IplImage)

createImage :: CvSize -> Depth -> Int -> IO (ErrCV IplImage)

Allocate memory for an IplImage with the given dimensions, number of color channels, and color depth.

cp_release_image :: FunPtr (Ptr IplImage -> IO ())

Release the memory allocated to an IplImage.

c_cvCloneImage :: Ptr Priv_IplImage -> IO (Ptr Priv_IplImage)

cloneImage :: IplImage -> IO (ErrCV IplImage)

c_get_size :: Ptr Priv_IplImage -> Ptr CvSize -> IO ()

c_cvCopy :: Ptr Priv_IplImage -> Ptr Priv_IplImage -> Ptr Priv_IplImage -> IO ()

copy :: IplImage -> IplImage -> IO ()

cvMerge :: Ptr Priv_IplImage -> Ptr Priv_IplImage -> Ptr Priv_IplImage -> Ptr Priv_IplImage -> Ptr Priv_IplImage -> IO ()

merge :: IplImage -> IplImage -> IplImage -> IplImage -> IplImage -> IO ()

wrap_getImageData :: Ptr Priv_IplImage -> IO (Ptr CUChar)

getImageData :: IplImage -> IO (ErrCV (Ptr CUChar))

getSize :: IplImage -> IO (ErrCV CvSize)

c_get_depth :: Ptr Priv_IplImage -> IO CInt

getDepth :: IplImage -> IO (ErrCV Depth)

c_get_nChannels :: Ptr Priv_IplImage -> IO CInt

getNumChannels :: Integral a => IplImage -> IO (ErrCV a)

wrap_getWidthStep :: Ptr Priv_IplImage -> IO CInt

getWidthStep :: IplImage -> IO (ErrCV Int)

cvConvertScale :: Ptr Priv_IplImage -> Ptr Priv_IplImage -> CDouble -> CDouble -> IO ()

convertScale :: IplImage -> IplImage -> Double -> Double -> IO ()

c_cvLoad :: CString -> Ptr Priv_CvMemStorage -> CString -> Ptr CString -> IO (Ptr a)

load :: String -> MemStorage -> Maybe String -> IO (ErrCV (ForeignPtr a, Maybe String))

cvGetSeqElem :: Ptr (Priv_CvSeq a) -> CInt -> IO (Ptr a)

seqNumElems :: Ptr (Priv_CvSeq a) -> IO CInt

seqToPList :: CvSeq a -> IO [ErrCV (ForeignPtr a)]

seqToList :: Storable a => CvSeq a -> IO [ErrCV a]

c_cvRectangle :: Ptr Priv_IplImage -> CInt -> CInt -> CInt -> CInt -> IO ()

rectangle :: IplImage -> CvRect -> IO ()

c_debug_print_image_header :: Ptr Priv_IplImage -> IO ()

Debugging function to print some of the internal details of an IplImage structure

mycheckPtr :: IO (Ptr a) -> IO (Ptr a)

Fail if an action results in a null pointer. NOTE: adapted from the allocated-processor package.

createForeignPtr :: FunPtr (Ptr a -> IO ()) -> IO (Ptr a) -> IO (ForeignPtr a)

Wrap a ForeignPtr around a Ptr after checking that the Ptr is non-null. The supplied finalizer is attached to the ForeignPtr. NOTE: adapted from the allocated-processor package.