HOpenCV.CV.HighGui

Description

While OpenCV was designed for use in full-scale applications and can be used within functionally rich UI frameworks (such as Qt*, WinForms*, or Cocoa*) or without any UI at all, sometimes there it is required to try functionality quickly and visualize the results. This is what the OpenCV.HighGui module has been designed for.

It provides easy interface to:

Create and manipulate windows that can display images and “remember” their content (no need to handle repaint events from OS).

Add trackbars to the windows, handle simple mouse events as well as keyboard commands.

Documentation

c_cvConvertImage :: Ptr Priv_IplImage -> Ptr Priv_IplImage -> CInt -> IO ()

convertImage :: IplImage -> IplImage -> Int -> IO ()

Function convertImage

List of functions in the module:

data Priv_CvCapture

type Capture = ForeignPtr Priv_CvCapture

c_cvCreateCameraCapture :: CInt -> IO (Ptr Priv_CvCapture)

pickAnyCam :: Int

self-documenting camera specification

cam :: Int -> Int

self-documenting camera specification

createCameraCapture :: Int -> IO (ErrCV Capture)

Open a capture stream from a connected camera. The parameter is the index of the camera to be used, or Nothing if it does not matter what camera is used. The returned action may be used to query for the next available frame.

c_cvCreateFileCapture :: CString -> IO (Ptr Priv_CvCapture)

createFileCapture :: String -> IO (ErrCV Capture)

cp_release_capture :: FunPtr (Ptr Priv_CvCapture -> IO ())

cv_release_capture :: Ptr Priv_CvCapture -> IO ()

release_capture :: Capture -> IO ()

c_cvQueryFrame :: Ptr Priv_CvCapture -> IO (Ptr Priv_IplImage)

queryFrame :: Capture -> IO (ErrCV IplImage)

If cvQueryFrame returns Nothing, try rewinding the video and querying again. If it still fails, raise an error. When a non-null frame is obtained, return it.

Windows

cvNamedWindow :: CString -> CInt -> IO CInt

type AutoSize = Bool

autoSize :: AutoSize

self-documenting window sizing specification

namedWindow :: String -> AutoSize -> IO ()

cvDestroyWindow :: CString -> IO ()

destroyWindow :: String -> IO ()

cvShowImage :: CString -> Ptr Priv_IplImage -> IO ()

showImage :: String -> IplImage -> IO ()

cvWaitKey :: CInt -> IO CInt

c_cvLoadImage :: CString -> CInt -> IO (Ptr Priv_IplImage)

waitKey :: Int -> IO (ErrCV Int)

loadImage :: String -> LoadImageColor -> IO (ErrCV IplImage)

newtype LoadImageColor

Determine the color model of an image loaded from a file.

Constructors

LoadImageColor

Fields

unLoadImageColor :: CInt

loadImageColor :: LoadImageColor

loadImageGrayscale :: LoadImageColor

c_cvSaveImage :: CString -> Ptr Priv_IplImage -> IO CInt

loadImageUnchanged :: LoadImageColor

saveImage :: String -> IplImage -> IO (ErrCV Int)

Trackbar

wrap_createTrackbar :: CString -> CString -> Ptr CInt -> CInt -> IO ()

createTrackbar :: String -> String -> Maybe Int -> Int -> IO ()

cvGetTrackbarPos :: CString -> CString -> IO CInt

getTrackbarPos :: String -> String -> IO (ErrCV Int)

cvSetTrackbarPos :: CString -> CString -> CInt -> IO ()

setTrackbarPos :: String -> String -> Int -> IO ()