HOpenCV.CV.Video

Description

Interfaces for grabbing images from cameras and video files, and for writing to video files.

Documentation

data Priv_CvVideoWriter

type CvVideoWriter = Ptr Priv_CvVideoWriter

wrap_cvCreateVideoWriter :: CString -> CInt -> CDouble -> CInt -> CInt -> IO (Ptr Priv_CvVideoWriter)

type FourCC = (Char, Char, Char, Char)

fourCC :: FourCC -> CInt

toFourCC :: String -> FourCC

Parse a four-character String into a FourCC code (e.g. XVID).

mpeg4CC :: FourCC

4-character code for MPEG-4.F

createVideoWriter :: String -> FourCC -> Double -> CvSize -> IO (ErrCV CvVideoWriter)

Create a video file writer. The parameters are the file name, the 4-character code (of the codec used to compress the frames (e.g. ('F','M','P','4') for MPEG-4), the framerate of the created video stream, and the size of the video frames. The returned action may be used to add frames to the video stream.

releaseVideoWriter :: CvVideoWriter -> IO ()

cvWriteFrame :: Ptr Priv_CvVideoWriter -> Ptr Priv_IplImage -> IO CInt

writeFrame :: CvVideoWriter -> IplImage -> IO (ErrCV Int)