HOpenCV.HCxCore

Description

Modulo abstracto para manejo de funciones de la libreria CxCore del core de OpenCV

Documentation

createMemStorageCV :: Int -> IO MemStorage

Function para crear un almacenamiento

createImageCV :: CvSize -> Depth -> Int -> IO IplImage

Function para crear una imagen

cloneImageCV :: IplImage -> IO IplImage

Function para clonar una imagen

getImageDataCV :: IplImage -> IO (Ptr CUChar)

Function para obtener datos de una imagen

getSizeCV :: IplImage -> IO CvSize

Function para obtener las medidas de una imagen

getDepthCV :: IplImage -> IO Depth

Function para obtener la profundidad una imagen

getNumChannelsCV :: Integral a => IplImage -> IO a

Function para obtener el numero de canales de una imagen

getWidthStepCV :: IplImage -> IO Int

Function para obtener el paso de ancho una imagen

loadCV :: String -> MemStorage -> Maybe String -> IO (ForeignPtr a, Maybe String)

Function para cargar una imagen de un almacenamiento

seqToPListCV :: CvSeq a -> IO [ForeignPtr a]

seqToListCV :: Storable a => CvSeq a -> IO [a]