HOpenCV.ImageProcessors

Description

ImageProcessors is a functional (Processor-based) interface to computer vision using Open

The Processor interface allows the primitives in this library to take care of all the allocation / deallocation of resources and other setup/teardown requirements, and to appropriately nest them when combining primitives.

Simple example:

win = window 0        -- The number is essentially a label for the window
cam = camera 0        -- Autodetect camera
edge = canny 30 190 3 -- Edge detecting processor using canny operator

test = cam >>> edge >>> win

The last expression is a processor that captures frames from camera and displays edge-detected version in the window.

Documentation

type ImageSink = IOSink Image

type ImageSource = IOSource () Image

type ImageProcessor = IOProcessor Image Image

type Image = IplImage

camera :: Int -> ImageSource

A capture device, using OpenCV's HighGui lib's cvCreateCameraCapture should work with most webcames. See OpenCV's docs for information. This processor outputs the latest image from the camera at each invocation.

videoFile :: String -> ImageSource

window :: Int -> ImageSink

A window that displays images. Note: windows with the same index will be the same window....is this ok? Response: Yes. It will display whatever images you give it.

namedWindowIP :: String -> Bool -> ImageSink

resizeIP :: Int -> Int -> InterpolationMethod -> ImageProcessor

OpenCV's cvResize

dilateIP :: Int -> ImageProcessor

OpenCV's cvDilate

cannyIP

Arguments

:: Int
Threshold 1

-> Int
Threshold 2

-> Int
Size

-> ImageProcessor

OpenCV's cvCanny

haarDetect

Arguments

:: String
Cascade filename (OpenCV comes with several, including ones for face detection)

-> Double
scale factor

-> Int
min neighbors

-> HaarDetectFlag
flags

-> CvSize
min size

-> IOProcessor Image [CvRect]

Wrapper for OpenCV's cvHaarDetectObjects and the surrounding required things (mem storage, cascade loading, etc).

drawRects :: IOProcessor (Image, [CvRect]) Image

OpenCV's cvRectangle, currently without width, color or line type control

runTill :: IOProcessor () b -> (b -> IO Bool) -> IO b

Runs the processor until a predicate is true, for predicates, and processors that take () as input (such as chains that start with a camera).

runTillKeyPressed :: Show a => IOProcessor () a -> IO ()

Name (and type) says it all.

keyPressed :: Show a => a -> IO Bool

Some general utility functions for use with Processors and OpenCV

Predicate for pressed keys